% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_acceptance_prob.R
\name{add_AP}
\alias{add_AP}
\title{Add Acceptance Probability (AP) to data frame}
\usage{
add_AP(data, M, pi)
}
\arguments{
\item{data}{The data frame containing all possible test plans.}

\item{M}{Monte Carlo simulation size.}

\item{pi}{A matrix containing the failure probability samples for different
test plans (rows represent samples, columns represent test plans).}
}
\value{
The data frame with an additional column for Acceptance Probability (AP) values.
}
\description{
This function calculates and adds the Acceptance Probability (AP) values to the input data frame.
}
