% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_expected_w.R
\name{add_expected_w}
\alias{add_expected_w}
\title{Add Expected Warranty Services Cost to Data Frame}
\usage{
add_expected_w(data, M, pi, cw, V, G)
}
\arguments{
\item{data}{A data frame containing test plan and cost information.}

\item{M}{Monte Carlo simulation size.}

\item{pi}{A matrix containing the failure probability samples for different test plans
(rows represent samples, columns represent test plans).}

\item{cw}{Average cost per warranty claim.}

\item{V}{Sales volume.}

\item{G}{Reliability growth cost.}
}
\value{
The data frame with an additional column for the expected warranty services cost (expW).
}
\description{
Calculate and add the expected warranty services (WS) cost based on the provided parameters
to the input data frame.
}
