% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_testing_cost.R
\name{add_testing_cost}
\alias{add_testing_cost}
\title{Add BRDT Testing Cost}
\usage{
add_testing_cost(data, cv, cf)
}
\arguments{
\item{data}{The data frame containing all possible test plans.}

\item{cv}{The average unit testing cost per testing sample.}

\item{cf}{The average fixed cost per RDT.}
}
\value{
The data frame with an additional column for Binomial RDT Testing Cost.
}
\description{
This function calculates and adds the Binomial RDT Testing Cost to the input data frame.
}
