% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacceptprob.R
\name{bacceptprob}
\alias{bacceptprob}
\title{Acceptance Probability for Binomial RDT}
\usage{
bacceptprob(M, n, c, pi)
}
\arguments{
\item{M}{Sample size of Monte Carlo simulation}

\item{n}{Testing sample size}

\item{c}{Maximum allowable failures}

\item{pi}{Failure probability}
}
\value{
Acceptance probability
}
\description{
Define the acceptance probability function which gets the probability of passing the test (for binomial RDT).
}
\examples{
pi <- pi_sim_beta(M = 100, seed = 10, a = 1, b = 1)
bacceptprob(M = 100, n = 10, c = 2, pi = pi)

}
