% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcost_WS.R
\name{bcost_WS}
\alias{bcost_WS}
\title{Calculate Warranty Services Cost}
\usage{
bcost_WS(cw, V, n, c, pi, M)
}
\arguments{
\item{cw}{Average cost per warranty claim}

\item{V}{Sales volume}

\item{n}{RDT sample size}

\item{c}{Maximum allowable failures}

\item{pi}{Failure probability}

\item{M}{Monte Carlo simulation size}
}
\value{
The warranty services cost
}
\description{
Define the cost function of warranty services (WS) after the decision of the test (for binomial RDT)
}
\examples{
pi = pi_sim_beta(M = 100, seed = 10, a = 1, b = 1)
cw = 200
V = 500
n = 100
c = 5
M = 1000

warranty_cost = bcost_WS(cw, V, n, c, pi, M)
warranty_cost

}
