% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_test_plans.R
\name{enum_test_plans}
\alias{enum_test_plans}
\title{Generate all test plans for BRDT (binary relibility demonstration test)}
\usage{
enum_test_plans(n, c, R)
}
\arguments{
\item{n}{Numeric vector, sample size values}

\item{c}{Numeric vector, maximum allowable failures values}

\item{R}{Numeric vector, lower level reliability requirement values}
}
\value{
A dataframe with all possible combinations of n, c, and R
}
\description{
Generates a dataframe including all possible combinations of sample size (n),
             maximum allowable failures (c), and lower level reliability requirement (R).
}
\examples{
n <- 1:10
c <- 0:5
R <- 0.8
enum_test_plans(n, c, R)

}
