% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mh_sampling_fail.R
\name{mh_sampling_fail}
\alias{mh_sampling_fail}
\title{Metropolis-Hastings Method for sampling the posterior distribution given Failing the RDT}
\usage{
mh_sampling_fail(
  sampleSize,
  startVal,
  sd,
  n,
  c,
  betaPriorParam1,
  betaPriorParam2
)
}
\arguments{
\item{sampleSize}{The size of the sample to be generated.}

\item{startVal}{The initial value of the Markov Chain.}

\item{sd}{The standard deviation for generating the proposal distribution.}

\item{n}{RDT testing sample size.}

\item{c}{The maximum number of allowable failures.}

\item{betaPriorParam1}{Parameter 1 for the Beta distribution.}

\item{betaPriorParam2}{Parameter 2 for the Beta distribution.}
}
\value{
A vector containing the Markov Chain sample of failure probabilities for the posterior given failing the RDT.
}
\description{
This function uses the Metropolis-Hastings algorithm to generate a sample of
failure probabilities given failinig the RDT.
}
\examples{
samples_fail <- mh_sampling_fail(sampleSize = 1000,
startVal = 0.2,
sd = 0.1,
n = 100,
c = 10,
betaPriorParam1 = 2,
betaPriorParam2 = 5)

samples_fail

}
