% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_stage_brdt.R
\name{multiStageRdt}
\alias{multiStageRdt}
\title{Multi-Stage Binomial Reliability Demonstration Test (MBRDT) Processor}
\usage{
multiStageRdt(
  maxStageNumber = 2,
  n = 1:20,
  c = 0:5,
  R = 0.85,
  seed = 123,
  M = 500,
  betaPriorParam1 = 1,
  betaPriorParam2 = 1,
  sampleSizePass = 200,
  sampleSizeFail = 200,
  startValForMH = 0.1,
  sdForMH = 1,
  cv = 25,
  cf = 0,
  cw = 6,
  V = 1000,
  G = 1e+05,
  burnInNum = 40,
  upperLimit = 1,
  growthRate = 0.7
)
}
\arguments{
\item{maxStageNumber}{Planning horizon, total number of stages}

\item{n}{Numeric vector, sample size values}

\item{c}{Numeric vector, maximum allowable failures values}

\item{R}{Numeric vector, lower level reliability requirement values}

\item{seed}{Random seed}

\item{M}{Simulation Sample Size}

\item{betaPriorParam1}{Parameter alpha of the prior beta distribution}

\item{betaPriorParam2}{Parameter beta of the prior beta distribution}

\item{sampleSizePass}{MH sampling size given passing the RDT}

\item{sampleSizeFail}{MH sampling size given failing the RDT}

\item{startValForMH}{Starting value for MH sampling}

\item{sdForMH}{Standard deviation of MH sampling}

\item{cv}{variable cost for test}

\item{cf}{fixed cost for test}

\item{cw}{warranty cost for each unit}

\item{V}{sales volume}

\item{G}{Reliability growth cost}

\item{burnInNum}{Burn-in number of MH sampling}

\item{upperLimit}{upper limit of reliability after reliability growth}

\item{growthRate}{parameter included in the RG model, larger value corresponds to slower growth}
}
\description{
This function performs a multi-stage reliability demonstration test (RDT) analysis.
It iteratively updates the parameters of the prior distribution and estimates the posterior distribution
of failure probabilities based on the test results at each stage.
}
