% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi_sim_beta.R
\name{pi_sim_beta}
\alias{pi_sim_beta}
\title{Monte Carlo Simulation for failure probabilities pi given Beta prior}
\usage{
pi_sim_beta(M, seed, a, b)
}
\arguments{
\item{M}{How many samples (pi values) you are going to simulate}

\item{seed}{Random seed for random sample}

\item{a}{Shape parameter 1 for beta distribution}

\item{b}{Shape parameter 2 for beta distribution}
}
\value{
The vector of failure probabilities simulated
}
\description{
Define the simulation function to generate failure probabilities
with Beta prior as conjugate prior to binomial distributions (for binomial RDT).
}
\examples{
pi <- pi_sim_beta(M = 100, seed = 10, a = 1, b = 1)
pi

}
