% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetFail.R
\name{targetFail}
\alias{targetFail}
\title{Target Distribution Given Failing the RDT}
\usage{
targetFail(x, n, c, betaPriorParam1, betaPriorParam2)
}
\arguments{
\item{x}{Failure probability (must be a value between 0 and 1).}

\item{n}{RDT testing sample size.}

\item{c}{The maximum number of allowable failures.}

\item{betaPriorParam1}{Parameter 1 for the Beta distribution.}

\item{betaPriorParam2}{Parameter 2 for the Beta distribution.}
}
\value{
The target distribution value.
}
\description{
This function calculates the target distribution given failing the RDT.
}
\examples{
target_fail <- targetFail(x = 0.2,
n = 100,
c = 10,
betaPriorParam1 = 2,
betaPriorParam2 = 5)

target_fail

}
